Change template code for UltraSonic Sensor with Soscon2017

+++ inc/resource_internal.h
@@ -48,7 +48,7 @@ struct _resource_s {
 };
 typedef struct _resource_s resource_s;
 
-typedef void (*resource_read_cb)(double value, void *data);
+typedef void (*resource_read_cb)(float value, void *data);
 
 struct _resource_read_cb_s {
 	resource_read_cb cb;
